Nvim image
Run my nvim config on other machines with this docker image.
Use the bin/nvim script to launch the image.
The bin script will make sure, if you're editing a file of a project,
the volume will be mapped to the top level of the project (defined by
git). If there's no git repo to be found, the current folder is mapped.